billdesksys/requirements.txt
dash==2.17.1
Flask==3.1.0
pandas==2.2.3
pywebview==5.3.2
Requests==2.32.3
waitress==3.0.2
